Extract timezonename into its own file
